+++ k8s/bindata.go
@@ -146,7 +146,7 @@ func resourcesCrdResourceJson() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "resources/crd-resource.json", size: 1941, mode: os.FileMode(420), modTime: time.Unix(1611143371, 0)}
+	info := bindataFileInfo{name: "resources/crd-resource.json", size: 1941, mode: os.FileMode(420), modTime: time.Unix(1611158407, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -266,7 +266,7 @@ func resourcesMockCrdTemplate() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "resources/mock-crd.template", size: 1159, mode: os.FileMode(420), modTime: time.Unix(1610528968, 0)}
+	info := bindataFileInfo{name: "resources/mock-crd.template", size: 1159, mode: os.FileMode(420), modTime: time.Unix(1611158407, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -286,7 +286,7 @@ func resourcesMockTemplate() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "resources/mock.template", size: 137, mode: os.FileMode(420), modTime: time.Unix(1610530253, 0)}
+	info := bindataFileInfo{name: "resources/mock.template", size: 137, mode: os.FileMode(420), modTime: time.Unix(1611158407, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -331,7 +331,7 @@ func resourcesSrvClusterRoleBindingJson() (*asset, error) {
 	return a, nil
 }
 
-var _resourcesSrvClusterRoleJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xdc\x90\x31\x4f\xc3\x30\x10\x85\xf7\xfe\x0a\xcb\x23\x0a\xad\xd8\x50\x57\x06\x76\x06\x16\xc4\x70\x71\x8e\x62\x35\xf1\x59\x77\xe7\x08\x81\xf2\xdf\x51\x12\x92\x2a\x28\xce\x50\x36\x46\xfb\x7b\xef\xe9\xde\xfb\xda\x19\x63\x8c\x85\xe8\x9f\x91\xc5\x53\xb0\x47\x63\xb9\x04\xb7\x87\xa4\xef\xc4\xfe\x13\xd4\x53\xd8\x9f\xef\x65\xef\xe9\xd0\xde\xd9\x62\x74\x9c\x7d\xa8\x7a\xed\x43\x9d\x44\x91\x9f\xa8\xc6\x09\x35\xa8\x50\x81\x82\x3d\x9a\x31\x7e\xf8\x0d\xd0\x60\x6f\xc0\x20\xd8\x94\x35\xde\x52\x44\x06\x25\xb6\x83\xa6\xfb\x31\x73\xaa\x51\xec\xd1\xbc\xcc\xce\x4b\xc6\x74\xea\x23\x53\x8a\x4b\xd1\x8c\xa7\x78\x92\xfe\x60\xbb\x10\xbc\x16\xcb\x28\x46\xa1\xc4\x0e\x33\x51\x6e\xac\x26\xb6\xc8\xb3\x83\x28\x68\x5a\x95\x5c\xd2\xb7\xe0\x46\x40\xa0\x6a\xdd\x3c\x80\xc9\xb8\xd9\xb0\x45\x2e\x33\xed\x6e\x7e\x39\xe7\x57\x57\x5c\x39\xfd\x9f\xc6\x8e\x54\xad\x76\xed\xff\x0f\xf8\x81\x6e\x0d\x0a\x72\xeb\x33\x0b\x63\xa8\x22\xf9\xa0\xeb\xb0\xc5\x0c\x71\x14\xde\xfc\xa9\x81\xf8\x6f\x96\x15\x74\x8c\x7a\x7d\x9d\x13\x6a\xb6\xd0\x6e\x7c\x77\xdf\x01\x00\x00\xff\xff\xc7\x14\x4b\x0e\x47\x04\x00\x00")
+var _resourcesSrvClusterRoleJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xdc\x90\x31\x4f\xc3\x30\x14\x84\xf7\xfe\x0a\xcb\x23\x0a\xad\xd8\x50\x57\x06\x76\x06\x16\xc4\xf0\xe2\x1c\xc5\x6a\xe2\x67\x3d\x3f\x47\x08\x94\xff\x8e\xd2\x90\x54\x41\x49\x87\xb2\x31\xda\x77\xf7\xe9\xee\x7d\x6d\x8c\x31\xc6\x52\xf4\xcf\x90\xe4\x39\xd8\xbd\xb1\x52\x92\xdb\x52\xd6\x77\x16\xff\x49\xea\x39\x6c\x8f\xf7\x69\xeb\x79\xd7\xde\xd9\x62\x48\x1c\x7d\xa8\x7a\xef\x43\x9d\x93\x42\x9e\xb8\xc6\x28\x35\x50\xaa\x48\xc9\xee\xcd\x80\x3f\xfd\x06\x6a\xd0\x07\x10\x12\x9a\xb2\xc6\x2d\x47\x08\x29\x8b\x3d\x79\xba\x9f\xb0\xe4\x1a\xc9\xee\xcd\xcb\x94\x3c\x33\xc6\xaa\x8f\xc2\x39\xce\x4d\x93\x3c\xe2\x39\xf5\x85\xed\xcc\xf0\x5a\xcc\x51\x82\xc4\x59\x1c\x56\x50\x6e\x98\x96\x6c\xb1\xae\xed\x92\x92\xe6\x45\xcb\x99\x7e\x49\x1c\x01\x17\x8b\xb6\x90\x72\xa5\xe4\xcd\xaf\xe4\xf4\xea\x8a\x2b\x2f\xf8\xa7\x9b\x45\xae\x16\xf7\xf6\xff\x3b\x7c\xc0\x2d\x89\x08\x55\x64\x1f\x74\x31\x89\x16\x2b\x8a\xe3\xf0\xe6\x0f\x0d\xc5\x7f\x73\xbd\x04\x27\xd0\xeb\xe7\x1c\xa0\xab\x83\x36\xc3\xbb\xfb\x0e\x00\x00\xff\xff\x40\x67\x87\x98\xf2\x03\x00\x00")
 
 func resourcesSrvClusterRoleJsonBytes() ([]byte, error) {
 	return bindataRead(
@@ -346,12 +346,12 @@ func resourcesSrvClusterRoleJson() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "resources/srv-cluster-role.json", size: 1095, mode: os.FileMode(420), modTime: time.Unix(1608931048, 0)}
+	info := bindataFileInfo{name: "resources/srv-cluster-role.json", size: 1010, mode: os.FileMode(420), modTime: time.Unix(1611328862, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
 
-var _resourcesSrvDeploymentJsonTemplate = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xac\x54\x4d\x6f\xd3\x40\x10\xbd\xf7\x57\x58\x3e\x07\x70\x0e\x5c\x72\x8b\x40\x48\x48\x50\xaa\x4a\xe5\x82\x38\x6c\x37\x4f\x61\xc5\xec\x07\x3b\xe3\xc2\x2a\xf2\x7f\x47\x6b\xb7\x49\x8c\xd7\x2e\x91\x9a\x53\x3c\xf3\xde\xcc\x7b\x33\x63\x1f\xae\xaa\xaa\xaa\x6a\x15\xcc\x57\x44\x36\xde\xd5\x9b\xfc\x14\xf8\xcd\xc3\xba\x5e\x0d\xc9\x9f\xc6\xed\x72\xf8\x3d\x02\xf9\x64\xe1\xe4\x29\x63\x21\x6a\xa7\x44\xd5\x9b\x6a\x28\xd4\x47\x49\xdd\x83\x78\x14\x7b\x6c\x12\x72\x19\x38\x86\xbd\x27\xbc\xf2\x01\x51\x89\x8f\xf5\x11\xd6\xad\x4e\x55\x9c\xb2\x58\x80\x3f\x42\x6b\x0e\xd0\xe3\xf6\x11\x81\x8c\x56\x59\xc0\xfa\xac\x1e\x83\xa0\x33\x7d\xa2\xcb\x2a\xd1\x3f\x3e\x95\x45\xff\xa7\xf0\x5e\x51\xd1\x06\x4b\x54\x82\x7d\x9a\xb6\x95\x14\x7a\x83\xb7\x9e\xc8\xb8\xfd\x5d\xd8\x29\x41\x79\x16\x02\x1b\x28\x67\xa7\xda\x4b\x1b\x78\x6e\x13\x17\x1a\x1b\x9b\xfb\x47\x5b\x55\x5c\xc2\x29\x83\xf8\x60\x34\xb6\x5a\xfb\xd6\xc9\xf5\xec\x4e\x57\x53\xaa\xf6\x4e\x94\x71\x88\xd9\xc0\xb7\xa2\x81\xb2\xad\x9e\x6d\xac\xda\xf7\xbd\x0e\x87\xd7\x1f\xf3\xff\xae\x2b\x34\x19\xc3\x6f\x5a\xa2\x1b\x4f\x46\xa7\x73\xe2\x29\xba\x5c\xe2\xe9\x60\xf3\x50\x17\x60\x0c\xdd\x46\x23\xe9\x9d\x77\x82\x3f\x32\xbb\x9c\x23\x21\x42\xed\xbe\x38\x4a\xb7\xde\xcb\x07\x43\xe0\xc4\x02\x5b\x6f\x2a\x89\x2d\xe6\x1b\x0d\xdc\xd6\x6d\xf9\xda\xbb\xcc\xbd\x80\x71\xc7\xc8\xef\xc9\xba\x69\x9a\x59\x74\xb7\xe0\x31\x82\x7d\x1b\x35\xe6\x4f\xef\x08\x25\x63\x8d\x3c\x8f\xeb\xb1\x3a\xb4\x79\xc0\xeb\xa6\xb1\x0b\x13\x3e\xc2\x2d\xac\x8f\xfd\x2a\xdf\x36\x9f\xcd\xf4\xa6\xcf\x7f\x0b\x6e\xaa\xc1\xd1\xaf\x16\x7c\xb1\xd0\x97\xd7\x39\xbf\x8f\x62\x66\x1a\xfd\x5e\x78\xd5\x22\x58\x54\x94\xd3\xed\x6f\xe9\xb7\x4a\x3c\xfb\x7d\x1b\xbe\xc0\x57\xdd\xdf\x00\x00\x00\xff\xff\xcf\x31\x12\x8a\x3e\x06\x00\x00")
+var _resourcesSrvDeploymentJsonTemplate = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xac\x56\x4d\x6f\xdb\x30\x0c\xbd\xf7\x57\x18\x3e\x37\x8d\x73\xd8\x25\xb7\x60\xc5\x80\x01\x6b\x17\x14\xe8\x2e\x43\x0f\x8c\xcc\xa5\xc2\xa8\x8f\x49\x54\x36\x23\xc8\x7f\x1f\x64\xe7\xcb\xb3\xec\x38\xc1\x72\x4a\xc4\x47\xf2\x3d\x3e\x53\xce\xf6\x2e\xcb\xb2\x2c\x07\x2b\xbf\xa1\xf3\xd2\xe8\x7c\x1e\x7f\x59\x3f\xdd\xcc\xf2\xfb\x26\xf8\x53\xea\x32\x1e\x3f\xa2\x25\x53\x29\xd4\x7c\x88\x28\x64\x28\x81\x21\x9f\x67\x4d\xa1\xfa\x94\x60\x85\xe4\x5b\x67\xfb\x26\x36\x96\x41\xed\x51\xad\x08\x27\xc6\xa2\x03\x36\x2e\x3f\xc2\x76\xf7\xa7\x2a\x1a\x14\x0e\xc0\xf7\xd0\xdc\x5b\x14\xed\xf6\x0e\x2d\x49\x01\x91\xc0\xec\xac\x9e\x47\x42\x11\xd3\x3b\xbc\x14\xb0\x78\xff\x92\x26\x3d\x92\x78\xcd\x28\x29\xc3\xb3\x03\xc6\x75\xd5\x6d\xcb\x95\xad\x05\xbe\x18\x22\xa9\xd7\xaf\xb6\x04\xc6\xf4\x2c\x18\x95\xa5\x18\xed\x72\x4f\x39\x70\xc9\x89\x2b\x85\xb5\xc5\xfd\xc3\x2d\x4b\x9a\x70\x8a\xa0\xdb\x48\x81\x0b\x21\x4c\xd0\xfc\xdc\xeb\xe9\x7d\x37\x55\x18\xcd\x20\x35\xba\x28\xe0\x7b\x52\x40\x5a\x56\x9d\x2d\x15\xac\xeb\x5e\xdb\xed\xc3\xe7\xf8\x7d\xb7\x4b\x34\x69\xc3\x97\x81\x68\x69\x48\x8a\xea\x3c\xf1\x74\x3a\x5c\xe2\xf0\xc0\xc6\xa1\x0e\xc0\x3c\x8a\xe0\x24\x57\x1f\x8d\x66\xfc\xc3\xbd\xe6\x1c\x13\x1c\x42\xf9\x55\x53\xf5\x62\x0c\x7f\x92\x84\xbe\xf2\x8c\x2a\x9f\x67\xec\x02\xf6\x37\x6a\x72\x83\x5e\xf8\x67\xa3\x63\xee\x15\x19\xaf\x1e\xe3\x9e\xcc\x8a\xa2\xe8\x45\xef\x06\x34\x3a\xf4\x26\x38\x81\xfd\x8f\xde\x11\x4a\x52\x49\xbe\x8c\xab\xb1\xc2\x86\x38\xe0\x59\x51\xa8\x81\x09\x1f\xe1\x0a\x95\x71\xb5\x95\x1f\x8a\x27\xd9\x7d\xa6\xcf\x3f\x03\x6a\xb2\x46\xd1\xaf\x80\xfe\x6a\xa2\xff\x9f\xe7\x4d\x7e\x80\x5b\xf7\x2f\xd1\x11\x15\x77\x15\x53\xab\xd8\x42\x4d\x26\x2b\x43\x5c\xae\x2e\xe1\xa6\xc1\xbb\x29\x19\x01\x34\x3d\xac\xfb\xd4\x33\x30\x3e\x94\xab\x7e\x8d\x6f\x03\x2a\x36\x86\x82\xc2\xa7\x78\x91\x5c\x56\x33\xc2\xa5\xc3\xc2\xd6\xac\x26\x4d\xf5\x51\x7e\x45\x06\x4b\xe0\xf7\x98\x9c\x90\x39\xa6\xc6\x61\xa9\xf3\x79\xf6\x03\xc8\xe3\x8d\xa6\xbf\x25\x23\x5d\x7c\x62\xac\xfb\x71\xde\x74\xb9\x5e\x39\xb9\x1c\x95\xe5\xea\x51\xd6\xef\xde\xb4\x98\x71\x94\x1d\x7a\x06\xc7\xa7\x3b\x7a\x41\xbf\xa1\xf2\xbd\xef\xe1\xe6\x9f\xc2\xdd\xee\x6f\x00\x00\x00\xff\xff\x9d\xf4\x96\xc0\xe6\x08\x00\x00")
 
 func resourcesSrvDeploymentJsonTemplateBytes() ([]byte, error) {
 	return bindataRead(
@@ -366,7 +366,7 @@ func resourcesSrvDeploymentJsonTemplate() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "resources/srv-deployment.json.template", size: 1598, mode: os.FileMode(420), modTime: time.Unix(1608931048, 0)}
+	info := bindataFileInfo{name: "resources/srv-deployment.json.template", size: 2278, mode: os.FileMode(420), modTime: time.Unix(1611329385, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
