+++ k8s/k8s.go
@@ -214,40 +214,6 @@ func (p *Provider) CreateResource(node *proto.Node) (*proto.Node, error) {
 	return node, nil
 }
 
-const (
-	// clustersURL is the k8s url for the cluster objects
-	clustersURL = "/apis/ensembleoss.io/v1/namespaces/default/clusters"
-
-	// resourcesURL is the k8s url for the resource objects
-	resourcesURL = "/apis/ensembleoss.io/v1/namespaces/default/resources"
-)
-
-func (p *Provider) TrackStuff(clt proto.EnsembleServiceClient) {
-
-	store := newStore()
-	newWatcher(store, p.client, clustersURL)
-	newWatcher(store, p.client, resourcesURL)
-
-	go func() {
-		task := store.pop(context.Background())
-		item := task.item
-
-		spec, err := decodeItem(item)
-		if err != nil {
-			panic(err)
-		}
-
-		c := &proto.Component{
-			Name:     item.Metadata.Name,
-			Spec:     spec,
-			Metadata: item.Metadata.Labels,
-		}
-		if _, err := clt.Apply(context.Background(), c); err != nil {
-			panic(err)
-		}
-	}()
-}
-
 func K8sFactory(logger hclog.Logger, c map[string]interface{}) (*Provider, error) {
 	config, err := GetConfig()
 	if err != nil {
