+++ k8s/kube_client.go
@@ -128,8 +128,7 @@ type Item struct {
 	Metadata *Metadata
 	Kind     string
 	Spec     map[string]interface{}
-	// Status   map[string]interface{}
-	Data map[string]interface{}
+	Data     map[string]interface{}
 }
 
 // InCluster returns whether we are running inside a Kubernetes pod
