+++ k8s/resources/srv-deployment.json.template
@@ -44,7 +44,25 @@
                                 "cpu": "10m",
                                 "memory": "50Mi"
                             }
-                        }
+                        },
+                        "args": [
+                            "server",
+                            "--boltdb",
+                            "/usr/local/ensemble/state.db"
+                        ],
+                        "volumeMounts": [
+                            {
+                                "name": "state-volume",
+                                "mountPath": "/usr/local/ensemble",
+                                "readOnly": false
+                            }
+                        ]
+                    }
+                ],
+                "volumes": [
+                    {
+                        "name": "state-volume",
+                        "emptyDir": {}
                     }
                 ],
                 "restartPolicy": "Always"
